# routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Todo List API",
    "version": "1.0.0",
    "title": "Todo list App"
  },
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/todo/edit/{id}": {
      "put": {
        "tags": [
          "Todo List"
        ],
        "summary": "Edit Todo List",
        "description": "",
        "operationId": "editTodo",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "List of user object",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/EditTodo"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
            
          },
          "404": {
            "description": "Failed"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/todo/delete": {
      "delete": {
        "tags": [
          "Todo List"
        ],
        "summary": "Delete Todo list",
        "description": "",
        "operationId": "deleteTodo",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "List of user object",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DeleteTodo"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
            
          },
          "404": {
            "description": "Failed"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "auth-token",
      "in": "header"
    }
  },
  "definitions": {
...
    "EditTodo":{
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "todo": {
          "type": "string"
        },
        "datetime": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "DeleteTodo":{
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
...
  }
}